feat: 앱 라우터에서 직접 참조하지 않고 페이지 레이어에서 조합
