+++ apps/web/src/_pages/setting/index.ts
@@ -1,2 +1,5 @@
-﻿export * from './SettingPage';
+export * from './SettingPage';
 export * from './MyFeedPage';
+export * from './notifications';
+export * from './password';
+export * from './privacy';
