+++ apps/web/src/_pages/setting/notifications/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingNotificationsPage';
