+++ apps/web/src/_pages/setting/password/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingPasswordPage';
