+++ apps/web/src/_pages/setting/privacy/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingPrivacyPage';
